requirements.txt
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Dropzone==1.6.0
Flask-WTF==1.0.1
googletrans==4.0.0rc1
gTTS==2.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyPDF2==2.12.0
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.4.0
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
